requirements.txt
streamlit
pandas
plotly
openai
pandasai
python-dotenv
matplotlib
